Switch to a simpler 'latest.json' scheme
Recently there was a change with the way that the Kano build servers
handled building of the Kano OS image, allowing building of 'flavour'
images but also providing opportunity for refactoring. During this
refactoring, the convention for naming of the generated the OS info JSON
changed to reflect the new ability to produce '.zip' files of the OS.
This fix resolves these troubles.

The 'latest.json' file on the server contains lots of information which
can be inferred from just the server location and the image name. This
file is currently being maintained manually whenever a new image is
released so reducing the number of changes required is preferential.
Thus in fixing the image naming convention issues, a switch to using a
simpler 'latest.json' file is implemented